# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Pet Management API
  description: Pet Management API with health logs, prescriptions, item master, and JWT authentication.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: '2'
servers:
  - url: http://localhost:9966/petmanagement/api/v2
...
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      tags: [user]
      operationId: addUser
      summary: Create a user (Admin Only)
      description: Creates a new user. This endpoint is protected and intended for admins. Use /auth/signup for public registration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    get:
      tags: [user]
      operationId: listUsers
      summary: Lists users with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: lastNameKana
          in: query
          required: false
          schema:
            type: string
        - name: firstNameKana
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
  /users/{userId}:
    get:
      tags: [user]
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags: [user]
      operationId: updateUser
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      tags: [user]
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
        403:
          description: Forbidden
...
  /prescriptions:
    get:
      tags: [prescription]
      operationId: listPrescriptions
      summary: List prescription master data with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionPage'
    post:
      tags: [prescription]
      operationId: addPrescription
      summary: Add new prescription master data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrescriptionFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
...
  /spaces:
    get:
      tags: [space]
      operationId: listSpaces
      summary: Get authenticated user's spaces
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Space'
    post:
      tags: [space]
      operationId: addSpace
      summary: Create a new space
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
  /spaces/{spaceId}/documents:
    get:
      tags: [document]
      operationId: listDocuments
      summary: List documents in a space
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
    post:
      tags: [document]
      operationId: addDocument
      summary: Create a new document
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
  /spaces/{spaceId}/documents/{documentId}:
    get:
      tags: [document]
      operationId: getDocument
      summary: Get a document
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    patch:
      tags: [document]
      operationId: updateDocument
      summary: Update a document partially
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpdateFields'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
    delete:
      tags: [document]
      operationId: deleteDocument
      summary: Delete a document
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index (0..N)
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: The size of the page to be returned
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
  schemas:
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          description: The size of the page
        totalElements:
          type: integer
          description: Total number of elements across all pages
        totalPages:
          type: integer
          description: Total number of pages
        number:
          type: integer
          description: Current page number (0-based)
      required:
        - size
        - totalElements
        - totalPages
        - number
    UserPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/UserResponse'
          required:
            - content
...
    PrescriptionPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Prescription'
          required:
            - content
...
    ProblemDetail:
      title: Problem Detail
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        timestamp:
          type: string
          format: date-time
    RoleName:
      type: string
      enum: [owner, vet, nurse, receptionist, clinic_admin, admin]
...
    PrescriptionCategory:
      type: string
      enum: [vaccine, heartworm, flea_tick, other]
...
    Role:
      title: Role
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          $ref: '#/components/schemas/RoleName'
      required:
        - id
        - name
...
    PrescriptionFields:
      title: Prescription Fields
      type: object
      properties:
        category:
          $ref: '#/components/schemas/PrescriptionCategory'
        name:
          type: string
          example: 'Heartgard Plus'
        form:
          type: string
          description: Tablet, Injection, Spot-on, etc.
          example: 'Tablet'
        strength:
          type: string
          example: 'S (up to 11kg)'
        note:
          type: string
      required:
        - category
        - name
    Prescription:
      allOf:
        - $ref: '#/components/schemas/PrescriptionFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - createdAt
            - updatedAt
...
    UserBase:
          type: object
          properties:
            username:
              type: string
              minLength: 1
              maxLength: 20
              example: john.doe
            firstName:
              type: string
              maxLength: 50
            lastName:
              type: string
              maxLength: 50
            firstNameKana:
              type: string
              maxLength: 50
              pattern: '^[ァ-ヶー]*$'
            lastNameKana:
              type: string
              maxLength: 50
              pattern: '^[ァ-ヶー]*$'
            email:
              type: string
              format: email
              maxLength: 255
            postalCode:
              type: string
              maxLength: 8
              pattern: '^$|^\d{3}-?\d{4}$'
            prefecture:
              type: string
              maxLength: 10
            city:
              type: string
              maxLength: 80
            address:
              type: string
              maxLength: 255
            telephone:
              type: string
              maxLength: 20
              pattern: '^$|^[0-9-]*$'
          required:
            - email
            - username
    UserRegistration:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            password:
              type: string
              minLength: 8
              maxLength: 128
              pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).+$'
              writeOnly: true
          required:
            - password
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            roles:
              type: array
              items:
                $ref: '#/components/schemas/Role'
            enabled:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - id
            - roles
            - enabled
            - createdAt
            - updatedAt
    UserUpdate:
      $ref: '#/components/schemas/UserBase'
...
    SpaceFields:
      title: Space Fields
      type: object
      properties:
        name:
          type: string
          example: 'My Space'
      required:
        - name
    Space:
      allOf:
        - $ref: '#/components/schemas/SpaceFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            userId:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - userId
            - createdAt
            - updatedAt
    DocumentFields:
      title: Document Fields
      type: object
      properties:
        parentDocId:
          type: string
          format: uuid
        title:
          type: string
          example: 'My Document'
        body:
          type: object
          additionalProperties: true
          description: JSONB body for block-based structure
      required:
        - title
    DocumentUpdateFields:
      title: Document Update Fields
      type: object
      properties:
        title:
          type: string
          example: 'Updated Document'
        body:
          type: object
          additionalProperties: true
          description: JSONB body for block-based structure
    Document:
      allOf:
        - $ref: '#/components/schemas/DocumentFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            spaceId:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - spaceId
            - createdAt
            - updatedAt
...